product_service/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
python-dotenv
djangorestframework-simplejwt